requirements.txt
streamlit==1.39.0
google-generativeai==0.7.1
python-dotenv==1.0.1